% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blues.note.R
\name{blues.note.plot}
\alias{blues.note.plot}
\title{Blues note plot}
\usage{
blues.note.plot(rink = FALSE, scale = NA, save = FALSE)
}
\arguments{
\item{rink}{boolean switch referencing the inclusion of the blues note in a rink plot (default is FALSE)}

\item{scale}{for plot generation}

\item{save}{boolean switch referencing an object save of blues.rink.plot.rds or blues.note.plot.rds (default is FALSE)}
}
\value{
blues.plot
}
\description{
Generates a plot with a blues note or a rink with a blues note.
}
\examples{
blues.note.plot()
}
