% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blues.note.R
\name{blues.note.processing}
\alias{blues.note.processing}
\title{Blues note processing}
\usage{
blues.note.processing(save = FALSE, ...)
}
\arguments{
\item{save}{boolean switch referencing a csv save of the intermediate blues.note.csv (default is FALSE)}

\item{...}{optional arguments for internal processing (unused)}
}
\value{
blues.note
}
\description{
Generates a frame used to plot a Blues note from simple information.
}
\examples{
blues.note.processing()
}
