% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blues.note.R
\name{rink.logo}
\alias{rink.logo}
\title{Rink logo}
\usage{
rink.logo(team)
}
\arguments{
\item{team}{teamName teamTriCode or teamId accepted (default is NA)}
}
\value{
team.logo
}
\description{
Rink logo
}
\examples{
rink.logo("STL")
}
