% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rink.R
\name{rink.plot}
\alias{rink.plot}
\title{Rink plot}
\usage{
rink.plot(save = FALSE)
}
\arguments{
\item{save}{boolean switch referencing an object save of rink.plot.rds (default is FALSE)}
}
\value{
rink.plot
}
\description{
Generates a plot containing an NHL regulation ice rink.
}
\examples{
rink.plot()
}
