% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rink.R
\name{rink.processing}
\alias{rink.processing}
\title{Rink processing}
\usage{
rink.processing(save = FALSE, ...)
}
\arguments{
\item{save}{boolean switch referencing a csv save of the intermediate rink.frame.csv (default is FALSE)}

\item{...}{optional arguments for internal processing (unused)}
}
\value{
rink.frame
}
\description{
Generates a frame used to plot an NHL regulation ice rink from simple information.
}
\examples{
rink.processing()
}
