% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rink.R
\name{rink}
\alias{rink}
\title{Rink}
\usage{
rink(team = NA)
}
\arguments{
\item{team}{teamName teamTriCode or teamId accepted (default is NA)}
}
\value{
rink.plot
}
\description{
Generates a plot containing an NHL regulation rink with a team logo.
}
\examples{
rink("STL")
}
