% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rink.R
\name{rink.save}
\alias{rink.save}
\title{Rink save}
\usage{
rink.save(rink, scale, file.name = "rink.png")
}
\arguments{
\item{rink}{provide either a rink function or rink plot}

\item{scale}{for plot generation}

\item{filename}{name of output file (default is rink.png)}
}
\value{
nothing
}
\description{
Wrapper to save a rink as a png
}
