% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rink.R
\name{rink.scale}
\alias{rink.scale}
\title{Rink Scale}
\usage{
rink.scale(unit = c("in", "mm", "cm", "ft", "px"), ...)
}
\arguments{
\item{unit}{desired unit to save plot c("in","mm","cm","ft","px")}

\item{...}{must include one of the following (height=#), (width=#), (scale=#)}
}
\value{
c(scale, height, width, unit) scale used to generate plot and the desired dimensions and unit to save the plot
}
\description{
Determine the scale to generate plot and dimensions to save.
}
\examples{
rink.scale("px", height=400, dpi=300)
rink.scale("mm", width=100)
rink.scale("in", scale=2)
}
